add multi-image encoder
