+++ hi-ml-multimodal/src/health_multimodal/common/__init__.py
@@ -10,5 +10,6 @@
 .. autosummary::
    :toctree:
 
+   device
    visualization
 """
