+++ hi-ml-multimodal/src/health_multimodal/image/model/__init__.py
@@ -3,9 +3,9 @@
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  -------------------------------------------------------------------------------------------
 
+from .encoder import ResnetType
 from .model import BaseImageModel
 from .model import ImageModel
-from .model import ResnetType
 from .model import get_biovil_resnet
 from .model import CXR_BERT_COMMIT_TAG
 from .model import BIOMED_VLP_CXR_BERT_SPECIALIZED
