+++ hi-ml-multimodal/src/health_multimodal/image/model/transformer.py
